refactor: enable additional compiler warnings and fix all resulting issues

+++ src/th_align.h
@@ -5,7 +5,7 @@
 
 #define TH_ALIGNOF(type) offsetof(struct { char c; type member; }, member)
 #define TH_ALIGNAS(align, ptr) ((void*)(((uintptr_t)(ptr) + ((align) - 1)) & ~((align) - 1)))
-#define TH_ALIGNUP(n, align) (((n) + (align) - 1) & ~((align) - 1))
+#define TH_ALIGNUP(n, align) (((n) + (size_t)(align) - 1) & ~((size_t)(align) - 1))
 #define TH_ALIGNDOWN(n, align) ((n) & ~((align) - 1))
 
 typedef long double th_max_align;
