+++ src/th_fcache.h
@@ -41,7 +41,7 @@ th_fcache_id_eq(th_fcache_id a, th_fcache_id b)
 TH_INLINE(uint32_t)
 th_fcache_id_hash(th_fcache_id id)
 {
-    return th_string_hash(id.path) + id.dir->fd;
+    return th_string_hash(id.path) + (uint32_t)id.dir->fd;
 }
 
 TH_DEFINE_HASHMAP(th_fcache_map, th_fcache_id, th_fcache_entry*, th_fcache_id_hash, th_fcache_id_eq, (th_fcache_id){0})
